refactor: consolidate TestHttpApplication and TestKafkaApplication into TestHttpController and TestKafkaListener
- Removed separate TestHttpApplication and TestKafkaApplication classes.
- Integrated their functionality into TestHttpController and TestKafkaListener respectively.
- Enhanced organization and structure of test classes for better maintainability.
