--- Jdempotent-spring-boot-postgres-starter/src/test/java/jdempotent/postgres/http/TestHttpApplication.java
@@ -1,11 +1,0 @@
-package jdempotent.postgres.http;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestHttpApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(TestHttpApplication.class, args);
-    }
-}
